Heavy commenting and some optimization
Commented most of the files, didn't get all the way through
LicenseParser. Renamed the TemporaryKey for the application. Made minor
edits to the preferences form to prepare for functionality allowing the
user to specify the location of the XML lookup file. Deleted some unused
code and variables in LicenseParser and Form1. Made optimizations to a
bit of the parsing code in LicenseParser. See comments for some areas
which need improvement.