Renamed build and added user-specified XML lookup functionality
Added exception handling for the "XML lookup not found" crash.
Coinciding with this, added a text bubble in the preferences window
where the user can specify the path manually (alternatively a prompt
appears if the XML is not found when Clean is pressed). Finally, renamed
build name from LicenseCleaner to Flexnet-License-Parser.